requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.64.0
thefuzz[speedup]>=0.22.0
unidecode>=1.3.0
openai>=1.0.0
supabase>=2.0.0
requests>=2.28.0
